+++ core/agents/brains/mat_brain.py
@@ -23,7 +23,7 @@ class MATBrain(AbstractBrain):
 
     @torch.no_grad()
     def get_action(self, state):
-        state = state.unsqueeze(0).to(self.device)
+        state = state.unsqueeze(0).float().to(self.device)
 
         q_values, attns = self.network.forward_attn(state)
         _, action = torch.max(q_values, dim=1)
